trekk kode for å finne personer fremstilt krav for ut i SøknadGrunnlagService så det gir mening å gjenbruke denne andre steder i koden
